requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
asyncpg==0.29.0
openai==1.40.0
pydantic==2.7.0
aiokafka==0.10.0
google-auth==2.30.0
google-api-python-client==2.130.0
twilio==9.1.0
pytest==8.2.0
pytest-asyncio==0.23.7
httpx==0.27.0
python-dotenv==1.0.1
mcp==1.0.0
